added a "tactical mode" on "Off -> 6C"
Tactical Mode is similar to Momentary Mode or Lockout Mode, but it has
three mode slots which are independent of other config values.  The default
is a classic tactical setup with "high, low, strobe", each with immediate
activation as soon as the button is pressed.

Each slot can be a regular ramp level 1 to 150, or a strobe-group mode.

To exit, press 6C or loosen/tighten the tailcap.

Aux LEDs follow the same pattern as Lockout Mode.

To configure, use 7H.  There are 3 options, one for each slot.
Enter a number 1 to 150 for a ramp mode, 0 for current strobe,
or 151+ for a specific strobe mode.