lockout mode: fixed manual memory timer and tint, added 3H next channel, and made 3H+ use mem level instead of lowest moon (this is needed for making the channel discernible, and also helps make aux LED controls stand out more)
The 3H mapping allows users to change channels without leaving lockout,
which was possible before with 3H to ramp tint, but doesn't work in new
multi-channel code because it uses 3C instead... and 3C is already used.
So the new 3H does sort of what the old 3H did, but in a way which is more
relevant to the new channel system.