add CHANNEL_USES_AUX to fix USE_AUX_RGB_LEDS_WHILE_ON bug
when running multichannel on a light without forward aux,
USE_AUX_RGB_LEDS_WHILE_ON is set by default. When switching to an aux
channel, the two conflict with unpredictable results. This can be seen
by building the previous commit for a D1 with aux channels enabled
(assign an actual integer value to USE_AUX_RGB_LEDS_WHILE_ON) as the
necessary channel definitions are inherited from the KR4 hwdef.