Add configurable RGB voltage aux with two vars, and moved config to battcheck menu compared to earlier implementation
This is a compromise between 1 var (not possible to fully configure) and
three (too many) and allows to confingure all posisbilities in the
previous implementation with 3.

Battcheck mode, 7H, items 3 and 4: 3 is threshold for RGB voltage aux to
switch to high level. 0 means "always high", and 151 "never high". In
addition, 152 disables the feature entirely (low *or* high). 4 is the
threshold below which RGB voltage aux will *always* be off, regardless
of high/low threshold setting, for lights where the aux are with the
main emitters and may interfere with moon mode.